chore: code tidy/organise
